fix(wework): unify workspace account and issue composer
