+++ wework/src/components/layout/DesktopSidebar.tsx
@@ -7,7 +7,6 @@ import {
   ChevronDown,
   ChevronRight,
   Columns2,
-  Download,
   Edit3,
   FolderOpen,
   FolderPlus,
@@ -26,7 +25,6 @@ import {
   Search,
   SquareTerminal,
   Sparkles,
-  UserRound,
   X,
 } from 'lucide-react'
 import {
@@ -80,7 +78,6 @@ import {
 import { useEscapeKey } from '@/hooks/useEscapeKey'
 import { useConfiguredKeybinding } from '@/hooks/useConfiguredKeybinding'
 import { useTranslation } from '@/hooks/useTranslation'
-import { getRuntimeConfig } from '@/config/runtime'
 import {
   canUseForProjectCreation,
   isCloudDevice,
@@ -138,7 +135,10 @@ import type {
   ArchiveRuntimeTaskResult,
 } from '@/features/workbench/workbenchContextTypes'
 import { getWorkbenchPaneKey } from './workbenchPaneIdentity'
-import { DesktopSettingsMenu } from './DesktopSettingsMenu'
+import {
+  DesktopSidebarAccount,
+  type DesktopSidebarAccountSettingsOptions,
+} from './DesktopSidebarAccount'
 import { DesktopWindowControls } from './DesktopWindowControls'
 import {
   DesktopSidebarHeader,
@@ -297,10 +297,7 @@ function getRuntimeTaskPinOverrideKey(deviceId: string, threadId: string) {
   return `${deviceId}\0${threadId}`
 }
 
-interface OpenSettingsOptions {
-  autoOpenAddCloudDeviceDialog?: boolean
-  settingsPage?: 'about' | 'connections'
-}
+type OpenSettingsOptions = DesktopSidebarAccountSettingsOptions
 
 type ProjectCreateMenuPosition = {
   top: number
@@ -353,17 +350,6 @@ const PROJECT_APPEARANCE_COLOR_VALUES: Record<string, string> = {
   red: '#ef4444',
   yellow: '#eab308',
 }
-function getSidebarAccountSummary(user: UserProfile | null, fallback: string) {
-  const userName = user?.user_name?.trim()
-  const email = user?.email?.trim()
-  const label = userName || email || fallback
-  const detail = email && email !== label ? email : fallback
-  return {
-    label,
-    detail,
-  }
-}
-
 function getStandaloneDeviceLabel(device: DeviceInfo): string {
   return device.name || device.device_id
 }
@@ -1247,117 +1233,6 @@ function GlobalImNotificationBell({
   )
 }
 
-function SidebarAppUpdateButton({ onBeforeInstall }: { onBeforeInstall?: () => void }) {
-  const { t } = useTranslation('common')
-  const appUpdate = useOptionalAppUpdate()
-  const buttonRef = useRef<HTMLButtonElement | null>(null)
-  const [errorTooltipPosition, setErrorTooltipPosition] = useState<{
-    left: number
-    top: number
-  } | null>(null)
-  const availableUpdate = appUpdate?.availableUpdate ?? null
-  const status = appUpdate?.status ?? 'idle'
-  const downloadProgress = appUpdate?.downloadProgress ?? null
-  const error = appUpdate?.error ?? null
-  const busy = status === 'checking' || status === 'installing'
-  const downloadPercent = downloadProgress
-    ? calculateSidebarUpdateDownloadPercent(
-        downloadProgress.downloadedBytes,
-        downloadProgress.totalBytes
-      )
-    : null
-
-  const showErrorTooltip = () => {
-    if (!error || !buttonRef.current) return
-    const rect = buttonRef.current.getBoundingClientRect()
-    setErrorTooltipPosition({
-      left: Math.min(rect.right + 8, Math.max(8, window.innerWidth - 268)),
-      top: Math.min(Math.max(8, rect.top + rect.height / 2), window.innerHeight - 8),
-    })
-  }
-
-  if (!appUpdate || !availableUpdate) return null
-
-  const title = availableUpdate
-    ? formatSidebarTemplate(
-        t('workbench.app_update_install', {
-          defaultValue: '更新到 {{version}}',
-          version: availableUpdate.version,
-        }),
-        { version: availableUpdate.version }
-      )
-    : t('workbench.app_update_check', '检查更新')
-  const downloadTitle =
-    downloadPercent === null
-      ? t('workbench.app_update_downloading', { defaultValue: '正在下载更新' })
-      : formatSidebarTemplate(
-          t('workbench.app_update_downloading_progress', {
-            defaultValue: '正在下载更新 {{progress}}%',
-            progress: downloadPercent,
-          }),
-          { progress: String(downloadPercent) }
-        )
-
-  return (
-    <div
-      className="group/update relative shrink-0"
-      onPointerEnter={showErrorTooltip}
-      onPointerLeave={() => setErrorTooltipPosition(null)}
-      onFocus={showErrorTooltip}
-      onBlur={() => setErrorTooltipPosition(null)}
-    >
-      <button
-        ref={buttonRef}
-        type="button"
-        data-testid="sidebar-app-update-button"
-        disabled={busy}
-        onClick={() => {
-          onBeforeInstall?.()
-          if (availableUpdate) {
-            void appUpdate.installUpdate()
-            return
-          }
-          void appUpdate.checkNow()
-        }}
-        title={error ?? (status === 'installing' ? downloadTitle : title)}
-        aria-label={error ?? (status === 'installing' ? downloadTitle : title)}
-        className={cn(
-          'group relative inline-flex h-8 w-8 shrink-0 items-center justify-center rounded-md transition-colors disabled:cursor-not-allowed disabled:opacity-60',
-          error
-            ? 'text-red-500 hover:bg-red-500/10'
-            : 'text-[rgb(var(--color-sidebar-text-secondary))] hover:bg-[rgb(var(--color-sidebar-hover))] hover:text-[rgb(var(--color-sidebar-text-primary))]'
-        )}
-      >
-        {status === 'installing' && downloadPercent !== null ? (
-          <SidebarUpdateDownloadProgress progress={downloadPercent} />
-        ) : busy ? (
-          <Loader2 className="h-4 w-4 animate-spin" />
-        ) : (
-          <Download className="sidebar-update-download-icon h-4 w-4" />
-        )}
-        {availableUpdate && !busy && (
-          <span className="absolute right-1 top-1 h-2 w-2 rounded-full bg-primary ring-2 ring-[rgb(var(--color-sidebar-hover))]" />
-        )}
-        {error && (
-          <span className="absolute right-1 top-1 h-2 w-2 rounded-full bg-red-500 ring-2 ring-[rgb(var(--color-sidebar-hover))]" />
-        )}
-      </button>
-      {error && errorTooltipPosition
-        ? createPortal(
-            <div
-              data-testid="sidebar-app-update-error"
-              style={errorTooltipPosition}
-              className="fixed z-system-popover w-[260px] -translate-y-1/2 rounded-lg border border-red-500/20 bg-popover px-3 py-2 text-xs font-medium leading-5 text-red-500 shadow-[0_12px_28px_rgba(0,0,0,0.18)] [overflow-wrap:anywhere]"
-            >
-              {error}
-            </div>,
-            document.body
-          )
-        : null}
-    </div>
-  )
-}
-
 function SidebarReleaseNotesCard({
   releaseNotes,
   onOpen,
@@ -1405,33 +1280,6 @@ function SidebarReleaseNotesCard({
         <X className="h-4 w-4" aria-hidden="true" />
       </button>
     </aside>
-  )
-}
-
-function calculateSidebarUpdateDownloadPercent(
-  downloadedBytes: number,
-  totalBytes: number | null
-): number | null {
-  if (!totalBytes || totalBytes <= 0) return null
-  return Math.min(100, Math.round((downloadedBytes / totalBytes) * 100))
-}
-
-function SidebarUpdateDownloadProgress({ progress }: { progress: number }) {
-  return (
-    <span
-      data-testid="sidebar-app-update-download-progress"
-      role="progressbar"
-      aria-label={`Update download ${progress}%`}
-      aria-valuemin={0}
-      aria-valuemax={100}
-      aria-valuenow={progress}
-      className="flex h-4 w-4 items-center justify-center rounded-full"
-      style={{
-        background: `conic-gradient(rgb(var(--color-primary)) ${progress}%, rgb(var(--color-sidebar-hover)) 0)`,
-      }}
-    >
-      <span className="h-2 w-2 rounded-full bg-[rgb(var(--color-sidebar))]" />
-    </span>
   )
 }
 
@@ -3135,26 +2983,12 @@ export function DesktopSidebar({
   const resizing = resizingProp ?? internalResizable.resizing
   const handleResizeStart = onResizeStartProp ?? internalResizable.handleResizeStart
   const showCloudConnectionEntry = isCloudConnectionUiAvailable()
-  const cloud = useOptionalCloudConnection()
-  const defaultWegentBackendUrl = getRuntimeConfig().wegentBackendUrl
-  const usesCloudAccount = showCloudConnectionEntry && Boolean(defaultWegentBackendUrl)
-  const requiresCloudLogin = usesCloudAccount && !cloud.isConnected
   const platform = getPlatform()
   const usesOverlayTitlebar = false
   const isWindowsTauri = isTauriRuntime() && platform === 'win'
   const appUpdate = useOptionalAppUpdate()
-  const hasAvailableAppUpdate = Boolean(appUpdate?.availableUpdate)
   const installedReleaseNotes = appUpdate?.installedReleaseNotes ?? null
   const experimentalFeaturesEnabled = useExperimentalFeaturesEnabled()
-  const sidebarAccount = requiresCloudLogin
-    ? {
-        label: t('workbench.account_cloud_title', 'Wegent 账户'),
-        detail: t('workbench.account_not_logged_in', '未登录'),
-      }
-    : getSidebarAccountSummary(
-        usesCloudAccount ? cloud.user : user,
-        t('workbench.account_fallback', '当前账号')
-      )
 
   const storageScope = getDesktopSidebarStorageScope(user)
   const projectsExpandedStorageKey = getDesktopSidebarStorageKey(storageScope, 'projectsExpanded')
@@ -3165,10 +2999,9 @@ export function DesktopSidebar({
     'expandedProjectIds'
   )
   const storageScopeRef = useRef(storageScope)
-  const [settingsMenuOpen, setSettingsMenuOpen] = useState(false)
   const [releaseNotesDialogOpen, setReleaseNotesDialogOpen] = useState(false)
-  const [accountCloudDialogOpen, setAccountCloudDialogOpen] = useState(false)
   const [imNotificationMenuOpen, setImNotificationMenuOpen] = useState(false)
+  const settingsMenuRef = useRef<HTMLDivElement>(null)
   const [archiveSectionMode, setArchiveSectionMode] = useState<'projects' | 'chats' | null>(null)
   const [forceArchiveSectionMode, setForceArchiveSectionMode] = useState<
     'projects' | 'chats' | null
@@ -3182,7 +3015,6 @@ export function DesktopSidebar({
   const [isArchivingProjectSection, setIsArchivingProjectSection] = useState(false)
   const [isArchivingChatSection, setIsArchivingChatSection] = useState(false)
   const [isArchivingPriority, setIsArchivingPriority] = useState(false)
-  const settingsMenuRef = useRef<HTMLDivElement>(null)
   const [projectCreateDialogOpen, setProjectCreateDialogOpen] = useState(false)
   const [standaloneWorkspaceDialogMode, setStandaloneWorkspaceDialogMode] =
     useState<StandaloneWorkspaceDialogMode | null>(null)
@@ -3705,6 +3537,23 @@ export function DesktopSidebar({
   }
 
   useEffect(() => {
+    if (!imNotificationMenuOpen) return
+
+    const handleOutsidePointer = (event: globalThis.MouseEvent | globalThis.PointerEvent) => {
+      if (!settingsMenuRef.current?.contains(event.target as Node)) {
+        setImNotificationMenuOpen(false)
+      }
+    }
+
+    document.addEventListener('pointerdown', handleOutsidePointer)
+    document.addEventListener('mousedown', handleOutsidePointer)
+    return () => {
+      document.removeEventListener('pointerdown', handleOutsidePointer)
+      document.removeEventListener('mousedown', handleOutsidePointer)
+    }
+  }, [imNotificationMenuOpen])
+
+  useEffect(() => {
     if (selectedRuntimeProjectId === null || !selectedRuntimeProjectAutoExpandKey) return
 
     const scopedAutoExpandKey = `${storageScope}:${selectedRuntimeProjectAutoExpandKey}`
@@ -3779,27 +3628,6 @@ export function DesktopSidebar({
     window.addEventListener('keydown', handleKeyDown)
     return () => window.removeEventListener('keydown', handleKeyDown)
   }, [priorityFilterShortcut, togglePriorityFilter])
-
-  useEffect(() => {
-    if (!settingsMenuOpen && !imNotificationMenuOpen) {
-      return
-    }
-
-    const handleOutsidePointer = (event: globalThis.MouseEvent | globalThis.PointerEvent) => {
-      if (!settingsMenuRef.current?.contains(event.target as Node)) {
-        setSettingsMenuOpen(false)
-        setImNotificationMenuOpen(false)
-      }
-    }
-
-    document.addEventListener('pointerdown', handleOutsidePointer)
-    document.addEventListener('mousedown', handleOutsidePointer)
-
-    return () => {
-      document.removeEventListener('pointerdown', handleOutsidePointer)
-      document.removeEventListener('mousedown', handleOutsidePointer)
-    }
-  }, [imNotificationMenuOpen, settingsMenuOpen])
 
   useEffect(() => {
     if (!currentRuntimeTask) return
@@ -4584,123 +4412,42 @@ export function DesktopSidebar({
             />
           )}
 
-          <div ref={settingsMenuRef} className="group/account relative shrink-0">
-            <div className="relative flex h-[60px] items-center rounded-[10px] transition-colors group-hover/account:bg-[rgb(var(--color-sidebar-hover))] group-focus-within/account:bg-[rgb(var(--color-sidebar-hover))]">
-              <button
-                type="button"
-                data-testid="settings-button"
-                onClick={() => {
-                  setImNotificationMenuOpen(false)
-                  setSettingsMenuOpen(open => !open)
-                }}
-                className={cn(
-                  'flex h-[60px] min-w-0 flex-1 items-center gap-3 rounded-[10px] py-2 pl-1.5 text-left text-[rgb(var(--color-sidebar-text-primary))]',
-                  hasAvailableAppUpdate ? 'pr-[72px]' : 'pr-10'
-                )}
-                title={t('workbench.account_and_settings', '账户与设置')}
-                aria-label={t('workbench.account_and_settings', '账户与设置')}
-                aria-expanded={settingsMenuOpen}
-              >
-                <span
-                  data-testid="sidebar-account-avatar"
-                  className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-primary/20 text-primary"
-                >
-                  <UserRound className="h-5 w-5" aria-hidden="true" />
-                </span>
-                <span className="min-w-0 flex-1">
-                  <span className="block truncate text-base font-semibold leading-[18px]">
-                    {sidebarAccount.label}
-                  </span>
-                  <span className="block truncate text-xs font-medium leading-4 text-[rgb(var(--color-sidebar-text-secondary))]">
-                    {sidebarAccount.detail}
-                  </span>
-                </span>
-              </button>
-              <div className="absolute right-1.5 top-1/2 flex -translate-y-1/2 items-center gap-0.5">
-                {hasAvailableAppUpdate && (
-                  <div data-testid="sidebar-app-update-action">
-                    <SidebarAppUpdateButton
-                      onBeforeInstall={() => {
-                        setSettingsMenuOpen(false)
-                        setImNotificationMenuOpen(false)
-                      }}
-                    />
-                  </div>
-                )}
-                {experimentalFeaturesEnabled && (
-                  <GlobalImNotificationBell
-                    devices={devices}
-                    imNotificationSettings={imNotificationSettings}
-                    menuOpen={imNotificationMenuOpen}
-                    menuContainerRef={settingsMenuRef}
-                    onMenuOpenChange={open => {
-                      if (open) setSettingsMenuOpen(false)
-                      setImNotificationMenuOpen(open)
-                    }}
-                    onToggleGlobalImNotification={onToggleGlobalImNotification}
-                    onOpenGlobalImNotificationSettings={onOpenGlobalImNotificationSettings}
-                    onOpenSettings={() => onOpenSettings()}
-                    onAddCloudDevice={() => {
-                      if (onOpenStandaloneFolderProject) {
-                        onOpenStandaloneFolderProject('remote', 'add-device')
-                      } else {
-                        setStandaloneRemoteDialogIntent('add-device')
-                        setStandaloneWorkspaceDialogMode('remote')
-                      }
-                    }}
-                  />
-                )}
-              </div>
-              {settingsMenuOpen && (
-                <DesktopSettingsMenu
-                  user={user}
-                  showLogout={usesCloudAccount ? cloud.isConnected : undefined}
-                  onOpenSettings={() => {
-                    setSettingsMenuOpen(false)
-                    onOpenSettings()
+          <DesktopSidebarAccount
+            user={user}
+            onOpenSettings={onOpenSettings}
+            onLogout={onLogout}
+            containerRef={settingsMenuRef}
+            trailingActions={
+              experimentalFeaturesEnabled ? (
+                <GlobalImNotificationBell
+                  devices={devices}
+                  imNotificationSettings={imNotificationSettings}
+                  menuOpen={imNotificationMenuOpen}
+                  menuContainerRef={settingsMenuRef}
+                  onMenuOpenChange={open => {
+                    setImNotificationMenuOpen(open)
                   }}
-                  onOpenAbout={() => {
-                    setSettingsMenuOpen(false)
-                    onOpenSettings({ settingsPage: 'about' })
-                  }}
-                  onLogin={
-                    requiresCloudLogin
-                      ? () => {
-                          setSettingsMenuOpen(false)
-                          setAccountCloudDialogOpen(true)
-                        }
-                      : undefined
-                  }
-                  onLogout={() => {
-                    setSettingsMenuOpen(false)
-                    if (usesCloudAccount) {
-                      cloud.disconnect()
-                      return
+                  onToggleGlobalImNotification={onToggleGlobalImNotification}
+                  onOpenGlobalImNotificationSettings={onOpenGlobalImNotificationSettings}
+                  onOpenSettings={() => onOpenSettings()}
+                  onAddCloudDevice={() => {
+                    if (onOpenStandaloneFolderProject) {
+                      onOpenStandaloneFolderProject('remote', 'add-device')
+                    } else {
+                      setStandaloneRemoteDialogIntent('add-device')
+                      setStandaloneWorkspaceDialogMode('remote')
                     }
-                    onLogout()
                   }}
                 />
-              )}
-            </div>
-          </div>
+              ) : null
+            }
+          />
 
           {installedReleaseNotes && (
             <AppReleaseNotesDialog
               open={releaseNotesDialogOpen}
               releaseNotes={installedReleaseNotes}
               onClose={() => setReleaseNotesDialogOpen(false)}
-            />
-          )}
-
-          {accountCloudDialogOpen && (
-            <CloudConnectionDialog
-              open
-              onlineCloudDeviceCount={0}
-              onClose={() => setAccountCloudDialogOpen(false)}
-              onOpenSettings={() => {
-                setAccountCloudDialogOpen(false)
-                onOpenSettings({ settingsPage: 'connections' })
-              }}
             />
           )}
