+++ wework/src/components/layout/DesktopWorkbenchLayout.tsx
@@ -15,6 +15,7 @@ import { buildRuntimeTaskRoute, isSettingsRoute, navigateTo } from '@/lib/naviga
 import { shouldUseNativeProjectDirectoryPicker } from '@/e2e/automation'
 import { cn } from '@/lib/utils'
 import { DesktopSidebar } from './DesktopSidebar'
+import type { DesktopSidebarAccountSettingsOptions } from './DesktopSidebarAccount'
 import { ProjectCreateDialog } from '@/components/projects/ProjectCreateDialog'
 import {
   StandaloneBlankProjectDialog,
@@ -402,6 +403,7 @@ export function DesktopWorkbenchLayout({ routeActive = true }: DesktopWorkbenchL
   const [sidebarPreviewOpen, setSidebarPreviewOpen] = useState(false)
   const [sidebarResizing, setSidebarResizing] = useState(false)
   const [settingsOpen, setSettingsOpen] = useState(() => isSettingsRoute(initialPath))
+  const settingsReturnPathRef = useRef(initialPath === '/todo' ? '/todo' : '/')
   const [autoOpenAddCloudDeviceDialog, setAutoOpenAddCloudDeviceDialog] = useState(false)
   const [blankProjectDialogOpen, setBlankProjectDialogOpen] = useState(false)
   const [standaloneWorkspaceDialogMode, setStandaloneWorkspaceDialogMode] =
@@ -545,10 +547,27 @@ export function DesktopWorkbenchLayout({ routeActive = true }: DesktopWorkbenchL
   )
 
   const openCloudDeviceSettings = useCallback(() => {
+    settingsReturnPathRef.current = '/'
     setAutoOpenAddCloudDeviceDialog(true)
     setSettingsOpen(true)
     navigateTo('/settings/connections')
   }, [])
+
+  const openSettings = useCallback(
+    (options: DesktopSidebarAccountSettingsOptions | undefined, returnPath: '/' | '/todo') => {
+      settingsReturnPathRef.current = returnPath
+      setAutoOpenAddCloudDeviceDialog(Boolean(options?.autoOpenAddCloudDeviceDialog))
+      setSettingsOpen(true)
+      navigateTo(
+        options?.autoOpenAddCloudDeviceDialog
+          ? '/settings/connections'
+          : options?.settingsPage
+            ? `/settings/${options.settingsPage}`
+            : '/settings'
+      )
+    },
+    []
+  )
 
   const openSidebarPreview = useCallback(() => {
     if (!effectiveSidebarCollapsed) return
@@ -864,17 +883,7 @@ export function DesktopWorkbenchLayout({ routeActive = true }: DesktopWorkbenchL
       onListDeviceDirectories={onListDeviceDirectories}
       onCreateDeviceDirectory={onCreateDeviceDirectory}
       projectSpaceApis={availableProjectSpaceApis}
-      onOpenSettings={options => {
-        setAutoOpenAddCloudDeviceDialog(Boolean(options?.autoOpenAddCloudDeviceDialog))
-        setSettingsOpen(true)
-        navigateTo(
-          options?.autoOpenAddCloudDeviceDialog
-            ? '/settings/connections'
-            : options?.settingsPage
-              ? `/settings/${options.settingsPage}`
-              : '/settings'
-        )
-      }}
+      onOpenSettings={options => openSettings(options, '/')}
       onLogout={onLogout}
     />
   )
@@ -925,9 +934,11 @@ export function DesktopWorkbenchLayout({ routeActive = true }: DesktopWorkbenchL
             onOpenRuntimeTask={onOpenRuntimeTask}
             onRefreshWorkLists={refreshWorkLists}
             onBack={() => {
+              const returnPath = settingsReturnPathRef.current
               setSettingsOpen(false)
               setAutoOpenAddCloudDeviceDialog(false)
-              navigateTo('/')
+              setCurrentPath(returnPath)
+              navigateTo(returnPath)
             }}
           />
         )}
@@ -940,6 +951,8 @@ export function DesktopWorkbenchLayout({ routeActive = true }: DesktopWorkbenchL
                 runtimeWork={state.runtimeWork}
                 services={services}
                 onOpenRuntimeTask={openProjectSpaceRuntimeTask}
+                onOpenSettings={options => openSettings(options, '/todo')}
+                onLogout={onLogout}
                 activeProjectRef={
                   workspaceTabs?.activeTab.kind === 'board'
                     ? boardRouteProjectRef(workspaceTabs.activeTab.contentRoute)
